delete element
Check the head pointer is NULL or not (List Empty). If it is not NULL then follow the next steps.
Read the element is going to be deleted.
By traversing the list upto temp is not NULL ,find the address of node that holds the element that matches to the element to be deleted and name it as temp. Also find the address of node before temp and name it as prev.
If prev pointer is NULL, link field of head node is assigned to the head and free the memory space of temp.
 If prev pointer is not NULL, link field of temp is assigned to link field of prev and free the memory of temp.
 If temp pointer is NULL, the element is not found.